fix: test folder moved to root dir
