+++ nightfall-client/package.json
@@ -2,10 +2,8 @@
   "name": "nightfall-client",
   "version": "1.9.0",
   "description": "A client for interacting with Nightfall",
-  "main": "index.mjs",
+  "main": "src/index.mjs",
   "scripts": {
-    "test": "npx mocha --timeout 0 --bail test/http.mjs test/neg-http.mjs",
-    "test-gas": "npx mocha --timeout 0 --bail test/tx-gas.mjs",
     "start": "node src/index.mjs",
     "dev": "nodemon src/index.mjs"
   },
@@ -39,14 +37,10 @@
     "safe-buffer": "^5.2.1",
     "web3": "^1.3.0",
     "winston": "^3.3.3",
-    "ws": "^7.4.3"
+    "ws": "^7.5.0"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "chai-http": "^4.3.0",
-    "mocha": "^8.1.3",
-    "nodemon": "^1.18.9",
-    "queue": "^6.0.2"
+    "nodemon": "^1.18.9"
   },
   "nodemonConfig": {
     "verbose": true,
