+++ package.json
@@ -4,8 +4,8 @@
   "description": "Optimistic Nightfall",
   "main": "./nightfall-client/src/index.mjs",
   "scripts": {
-    "test": "npx mocha --timeout 0 --bail nightfall-client/test/http.mjs nightfall-client/test/neg-http.mjs",
-    "test-gas": "npx mocha --timeout 0 --bail nightfall-client/test/tx-gas.mjs",
+    "test": "mocha --timeout 0 --bail test/http.mjs test/neg-http.mjs",
+    "test-gas": "mocha --timeout 0 --bail test/tx-gas.mjs",
     "lint": "eslint . --ext js,mjs",
     "prepare": "husky install"
   },
@@ -22,11 +22,19 @@
   "devDependencies": {
     "@commitlint/cli": "^12.1.1",
     "@commitlint/config-conventional": "^12.1.1",
+    "axios": "^0.21.1",
+    "chai": "^4.3.4",
+    "chai-as-promised": "^7.1.1",
+    "chai-http": "^4.3.0",
     "eslint": "^7.25.0",
     "eslint-config-codfish": "^10.0.0",
+    "general-number": "^1.0.1",
     "husky": "^6.0.0",
     "mocha": "^8.3.2",
     "prettier": "^2.2.1",
-    "prettier-plugin-solidity": "^1.0.0-beta.10"
+    "prettier-plugin-solidity": "^1.0.0-beta.10",
+    "queue": "^6.0.2",
+    "web3": "^1.3.6",
+    "ws": "^7.5.0"
   }
 }
