+++ setup-nightfall
@@ -3,9 +3,4 @@
 # Install node dependencies 
 npm ci
 
-# Install node dependencies of nightfall-client
-pushd nightfall-client
-npm ci
-popd
-
 docker-compose build --build-arg GPR_TOKEN=${GPR_TOKEN} --no-cache
