+++ test/http.mjs
@@ -6,7 +6,7 @@ import chaiAsPromised from 'chai-as-promised';
 import Queue from 'queue';
 import gen from 'general-number';
 import WebSocket from 'ws';
-import sha256 from '../src/utils/crypto/sha256.mjs';
+import sha256 from '../nightfall-client/src/utils/crypto/sha256.mjs';
 import {
   closeWeb3Connection,
   submitTransaction,
